[refactor]organize folders
