+++ Deadlock.uproject
@@ -7,7 +7,10 @@
 		{
 			"Name": "Deadlock",
 			"Type": "Runtime",
-			"LoadingPhase": "Default"
+			"LoadingPhase": "Default",
+			"AdditionalDependencies": [
+				"Engine"
+			]
 		}
 	],
 	"Plugins": [
