+++ Source/Deadlock/GameMode/DeadlockGameState.h
@@ -1,0 +1,17 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/GameStateBase.h"
+#include "DeadlockGameState.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class DEADLOCK_API ADeadlockGameState : public AGameStateBase
+{
+	GENERATED_BODY()
+	
+};
