+++ Source/Deadlock/GameMode/DeadlockPlayerController.h
@@ -1,0 +1,17 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/PlayerController.h"
+#include "DeadlockPlayerController.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class DEADLOCK_API ADeadlockPlayerController : public APlayerController
+{
+	GENERATED_BODY()
+	
+};
