+++ Source/Deadlock/GameMode/DeadlockPlayerState.h
@@ -1,0 +1,17 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/PlayerState.h"
+#include "DeadlockPlayerState.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class DEADLOCK_API ADeadlockPlayerState : public APlayerState
+{
+	GENERATED_BODY()
+	
+};
